% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndonesiaStemmeR.R
\name{stemKalimat}
\alias{stemKalimat}
\title{Word stemmer for Bahasa Indonesia}
\usage{
stemKalimat(kalimat)
}
\arguments{
\item{kalimat}{}
}
\description{
Provides function to stem (retrieve root word of a word) a word in
Bahasa Indonesia using Nazief and Andriani's algorithm.
It consists set of functions to remove prefixes, suffixes or both.
This package is made based on katadasaR package
}
\examples{
stemKalimat("percobaan")

}
\author{
Aisyah Zakiah
}
